// src/types/enums.ts
export enum MealType {
  Breakfast = 'Breakfast',
  Lunch = 'Lunch',
  Dinner = 'Dinner',
  Both = 'Both',
  None = 'None'
}

export enum FoodPreference {
  Veg = 'Veg',
  NonVeg = 'Non-Veg'
}

export enum PlanType {
  Regular = 'Regular',
  Trial = 'Trial',
  Monthly20 = 'Monthly 20',
  Daily = 'Daily'
}

export enum OrderStatus {
  Booked = 'Booked',
  Dispatched = 'Dispatched',
  Delivered = 'Delivered',
  Cancelled = 'Cancelled'
}

export enum TransactionType {
  Credit = 'Credit',
  Debit = 'Debit'
}